requirements.txt
customtkinter>=5.2.0
pyinstaller>=6.0
Pillow>=10.0
